+++ cli/connhelper/commandconn/commandconn_linux.go
@@ -1,4 +1,4 @@
-package connhelper
+package commandconn
 
 import (
 	"os/exec"
