+++ cli/connhelper/commandconn/commandconn_nolinux.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package connhelper
+package commandconn
 
 import (
 	"os/exec"
