+++ cli/connhelper/ssh/ssh.go
@@ -1,5 +1,4 @@
 // Package ssh provides the connection helper for ssh:// URL.
-// Requires Docker 18.09 or later on the remote host.
 package ssh
 
 import (
@@ -8,16 +7,8 @@ import (
 	"github.com/pkg/errors"
 )
 
-// New returns cmd and its args
-func New(daemonURL string) (string, []string, error) {
-	sp, err := parseSSHURL(daemonURL)
-	if err != nil {
-		return "", nil, errors.Wrap(err, "SSH host connection is not valid")
-	}
-	return "ssh", append(sp.Args(), []string{"--", "docker", "system", "dial-stdio"}...), nil
-}
-
-func parseSSHURL(daemonURL string) (*sshSpec, error) {
+// ParseURL parses URL
+func ParseURL(daemonURL string) (*Spec, error) {
 	u, err := url.Parse(daemonURL)
 	if err != nil {
 		return nil, err
@@ -26,19 +17,19 @@ func parseSSHURL(daemonURL string) (*sshSpec, error) {
 		return nil, errors.Errorf("expected scheme ssh, got %q", u.Scheme)
 	}
 
-	var sp sshSpec
+	var sp Spec
 
 	if u.User != nil {
-		sp.user = u.User.Username()
+		sp.User = u.User.Username()
 		if _, ok := u.User.Password(); ok {
 			return nil, errors.New("plain-text password is not supported")
 		}
 	}
-	sp.host = u.Hostname()
-	if sp.host == "" {
+	sp.Host = u.Hostname()
+	if sp.Host == "" {
 		return nil, errors.Errorf("no host specified")
 	}
-	sp.port = u.Port()
+	sp.Port = u.Port()
 	if u.Path != "" {
 		return nil, errors.Errorf("extra path after the host: %q", u.Path)
 	}
@@ -51,20 +42,22 @@ func parseSSHURL(daemonURL string) (*sshSpec, error) {
 	return &sp, err
 }
 
-type sshSpec struct {
-	user string
-	host string
-	port string
+// Spec of SSH URL
+type Spec struct {
+	User string
+	Host string
+	Port string
 }
 
-func (sp *sshSpec) Args() []string {
+// Args returns args except "ssh" itself and "-- ..."
+func (sp *Spec) Args() []string {
 	var args []string
-	if sp.user != "" {
-		args = append(args, "-l", sp.user)
+	if sp.User != "" {
+		args = append(args, "-l", sp.User)
 	}
-	if sp.port != "" {
-		args = append(args, "-p", sp.port)
+	if sp.Port != "" {
+		args = append(args, "-p", sp.Port)
 	}
-	args = append(args, sp.host)
+	args = append(args, sp.Host)
 	return args
 }
